feat: Web Experiment - refactor script for package
