+++ packages/experiment-tag/rollup.config.js
@@ -52,8 +52,8 @@ const getCommonBrowserConfig = (target) => ({
 const getOutputConfig = (outputOptions) => ({
   output: {
     dir: 'dist',
-    name: 'Experiment-Tag',
-    banner: `/* ${packageJson.name} v${packageJson.version} - For license info see https://unpkg.com/@amplitude/experiment-tag@${packageJson.version}/LICENSES */`,
+    name: 'WebExperiment',
+    banner: `/* ${packageJson.name} v${packageJson.version} - For license info see https://unpkg.com/@amplitude/experiment-tag@${packageJson.version}/files/LICENSE */`,
     ...outputOptions,
   },
 });
