+++ packages/experiment-tag/src/index.ts
@@ -1,23 +1,29 @@
 import { DefaultWebExperimentClient } from './experiment';
+import { WebExperimentConfig } from './types';
 
-const API_KEY = '{{DEPLOYMENT_KEY}}';
-const initialFlags = '{{INITIAL_FLAGS}}';
-const serverZone = '{{SERVER_ZONE}}';
-const pageObjects = '{{PAGE_OBJECTS}}';
+export const initialize = (
+  apiKey: string,
+  initialFlags: string,
+  pageObjects: string,
+  config: WebExperimentConfig,
+): void => {
+  DefaultWebExperimentClient.getInstance(
+    apiKey,
+    initialFlags,
+    pageObjects,
+    config,
+  )
+    .start()
+    .finally(() => {
+      // Remove anti-flicker css if it exists
+      document.getElementById('amp-exp-css')?.remove();
+    });
+};
 
-DefaultWebExperimentClient.getInstance(API_KEY, initialFlags, pageObjects, {
-  serverZone: serverZone,
-})
-  .start()
-  .then(() => {
-    // Remove anti-flicker css if it exists
-    document.getElementById('amp-exp-css')?.remove();
-  });
-
-export { WebExperimentClient } from 'web-experiment';
-export { WebExperimentConfig } from 'config';
 export {
   ApplyVariantsOptions,
   RevertVariantsOptions,
   PreviewVariantsOptions,
+  WebExperimentClient,
+  WebExperimentConfig,
 } from 'types';
