+++ packages/experiment-tag/src/types.ts
@@ -1,4 +1,6 @@
 import { EvaluationCondition } from '@amplitude/experiment-core';
+import { ExperimentConfig } from '@amplitude/experiment-js-client';
+import { ExperimentClient, Variants } from '@amplitude/experiment-js-client';
 
 import { MessageType } from './message-bus';
 
@@ -32,3 +34,34 @@ export type PageObject = {
 };
 
 export type PageObjects = { [flagKey: string]: { [id: string]: PageObject } };
+
+export interface WebExperimentConfig extends ExperimentConfig {
+  /**
+   * Determines whether the default implementation for handling navigation {@link setDefaultUrlChangeHandler} will be used
+   * If this is set to false, for single-page applications:
+   * 1. The variant actions applied will be based on the context (user, page URL) when the web experiment script was loaded
+   * 2. Custom handling of navigation should be implemented such that variant actions applied on the site reflect the latest context
+   */
+  useDefaultNavigationHandler?: boolean;
+}
+
+export const Defaults: WebExperimentConfig = {
+  useDefaultNavigationHandler: true,
+};
+
+
+/**
+ * Interface for the Web Experiment client.
+ */
+
+export interface WebExperimentClient {
+  start(): void;
+  getExperimentClient(): ExperimentClient;
+  applyVariants(options?: ApplyVariantsOptions): void;
+  revertVariants(options?: RevertVariantsOptions): void;
+  previewVariants(options: PreviewVariantsOptions): void;
+  getVariants(): Variants;
+  getActiveExperiments(): string[];
+  getActivePages(): PageObjects;
+  setRedirectHandler(handler: (url: string) => void): void;
+}
